@@ -74,7 +74,7 @@
     "rollup-plugin-babel": "^4.3.3",
     "rollup-plugin-typescript2": "^0.24.3",
     "tslib": "1.10.0",
-    "typescript": "^3.6.3"
+    "typescript": "3.7.4"
   },
   "bundlesize": [
     {
